move the `load_program_texts`, `get_text`, `log_text` to a separate module `io.logger.py`
