+++ src/dfastbe/bank_erosion/bank_erosion.py
@@ -55,10 +55,8 @@ from dfastbe.bank_erosion.data_models import (
 )
 from dfastbe.bank_erosion.debugger import Debugger
 from dfastbe.bank_erosion.utils import BankLinesProcessor, intersect_line_mesh, write_km_eroded_volumes
-from dfastbe.io.config import (
-    ConfigFile,
-    log_text,
-)
+from dfastbe.io.config import ConfigFile
+from dfastbe.io.logger import log_text
 from dfastbe.io.data_models import LineGeometry
 from dfastbe.kernel import get_zoom_extends
 from dfastbe.bank_erosion.utils import (
