+++ src/dfastbe/io/config.py
@@ -39,9 +39,7 @@ from geopandas.geodataframe import GeoDataFrame
 from geopandas.geoseries import GeoSeries
 from shapely.geometry import LineString
 from dfastbe.io.file_utils import absolute_path, relative_path
-
-
-PROGTEXTS: Dict[str, List[str]]
+from dfastbe.io.logger import log_text
 
 
 class ConfigFile:
@@ -1174,42 +1172,6 @@ def get_bbox(
 
     return bbox
 
-def load_program_texts(file_name: Union[str, Path]) -> None:
-    """Load texts from a configuration file, and store globally for access.
-
-    This routine reads the text file "file_name", and detects the keywords
-    indicated by lines starting with [ and ending with ]. The content is
-    placed in a global dictionary PROGTEXTS which may be queried using the
-    routine "get_text". These routines are used to implement multi-language support.
-
-    Arguments
-    ---------
-    file_name : str
-        The name of the file to be read and parsed.
-    """
-    global PROGTEXTS
-
-    all_lines = open(file_name, "r").read().splitlines()
-    data: Dict[str, List[str]] = {}
-    text: List[str] = []
-    key = None
-    for line in all_lines:
-        r_line = line.strip()
-        if r_line.startswith("[") and r_line.endswith("]"):
-            if key is not None:
-                data[key] = text
-            key = r_line[1:-1]
-            text = []
-        else:
-            text.append(line)
-    if key in data.keys():
-        raise ValueError(f"Duplicate entry for {key} in {file_name}.")
-
-    if key is not None:
-        data[key] = text
-
-    PROGTEXTS = data
-
 
 def _move_parameter_location(
     config: ConfigParser,
@@ -1346,69 +1308,3 @@ class SimulationFilesError(Exception):
     """Custom exception for configuration file errors."""
 
     pass
-
-def get_text(key: str) -> List[str]:
-    """
-    Query the global dictionary of texts via a string key.
-
-    Query the global dictionary PROGTEXTS by means of a string key and return
-    the list of strings contained in the dictionary. If the dictionary doesn't
-    include the key, a default string is returned.
-
-    Parameters
-    ----------
-    key : str
-        The key string used to query the dictionary.
-
-    Returns
-    -------
-    text : List[str]
-        The list of strings returned contain the text stored in the dictionary
-        for the key. If the key isn't available in the dictionary, the routine
-        returns the default string "No message found for <key>"
-    """
-
-    global PROGTEXTS
-
-    try:
-        str_value = PROGTEXTS[key]
-    except:
-        str_value = ["No message found for " + key]
-    return str_value
-
-def log_text(
-        key: str,
-        file: Optional[TextIO] = None,
-        data: Dict[str, Any] = {},
-        repeat: int = 1,
-        indent: str = "",
-) -> None:
-    """
-    Write a text to standard out or file.
-
-    Arguments
-    ---------
-    key : str
-        The key for the text to show to the user.
-    file : Optional[TextIO]
-        The file to write to (None for writing to standard out).
-    data : Dict[str, Any]
-        A dictionary used for placeholder expansions (default empty).
-    repeat : int
-        The number of times that the same text should be repeated (default 1).
-    indent : str
-        String to use for each line as indentation (default empty).
-
-    Returns
-    -------
-    None
-    """
-    str_value = get_text(key)
-    for _ in range(repeat):
-        for s in str_value:
-            sexp = s.format(**data)
-            if file is None:
-                print(indent + sexp)
-            else:
-                file.write(indent + sexp + "\n")
-
